♻️ Extract Finance endpoint host into FinanceEndpointHost
Mirrors the Catalog extraction so the same Register/InitializeDatabaseAsync
shape can be called from both standalone Program.cs and AllInOne.
