+++ src/Finance.Endpoint/Program.cs
@@ -1,41 +1,12 @@
-using Finance.Data;
-using Finance.Data.Seed;
-using Finance.Endpoint.Handlers;
-using ITOps.Shared.EndpointConfiguration;
-using ITOps.Shared.Sqlite;
-using Microsoft.EntityFrameworkCore;
+using Finance.Endpoint;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-const string EndpointName = "Finance";
-
-HostApplicationBuilder hostBuilder = Host.CreateApplicationBuilder(args);
-
-var sqliteConnectionString = SqliteStorage.GetConnectionString("finance");
-
-hostBuilder.Services.AddDbContext<FinanceDbContext>(options =>
-    options.UseSqlite(sqliteConnectionString));
-
-// Configure NServiceBus. Handlers are registered explicitly via the
-// source-generated AddHandler<T>() path (NServiceBus 10.2); reflection
-// scanning still finds them too and the registrations dedupe.
-var endpointConfiguration = new EndpointConfiguration(EndpointName);
-endpointConfiguration.Configure(sqliteConnectionString);
-endpointConfiguration.AddHandler<OrderCancelledHandler>();
-endpointConfiguration.AddHandler<OrderPlacedHandler>();
-endpointConfiguration.AddHandler<SubmitBillingAddressHandler>();
-endpointConfiguration.AddHandler<SubmitDeliveryOptionHandler>();
-endpointConfiguration.AddHandler<SubmitOrderItemsHandler>();
-hostBuilder.Services.AddNServiceBusEndpoint(endpointConfiguration);
+var hostBuilder = Host.CreateApplicationBuilder(args);
+FinanceEndpointHost.Register(hostBuilder);
 
 var host = hostBuilder.Build();
-var hostEnvironment = host.Services.GetRequiredService<IHostEnvironment>();
-Console.Title = hostEnvironment.ApplicationName;
+Console.Title = host.Services.GetRequiredService<IHostEnvironment>().ApplicationName;
 
-using (var scope = host.Services.CreateScope())
-{
-    var dbContext = scope.ServiceProvider.GetRequiredService<FinanceDbContext>();
-    await DatabaseInitializer.InitializeAsync(dbContext);
-}
-
+await FinanceEndpointHost.InitializeDatabaseAsync(host);
 await host.RunAsync();
